requirements.txt
anyio==4.8.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
exceptiongroup==1.2.2
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
python-dotenv==1.0.1
python-telegram-bot==21.11.1
sniffio==1.3.1
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
